Add ability to show a limited view of the active extensions page to authenticated users. Control for the call is only given to extensions that are assigned to that users.
git-svn-id: http://fusionpbx.googlecode.com/svn/trunk@290 e2954ef0-caed-11de-bd98-e341d93cc3cc

+++ fusionpbx/mod/calls_active/v_calls_active_extensions_inc.php
@@ -17,8 +17,8 @@
 
 	The Initial Developer of the Original Code is
 	Mark J Crane <markjcrane@fusionpbx.com>
-	Portions created by the Initial Developer are Copyright (C) 2008-2010
-	the Initial Developer. All Rights Reserved.
+	Copyright (C) 2010
+	All Rights Reserved.
 
 	Contributor(s):
 	Mark J Crane <markjcrane@fusionpbx.com>
@@ -26,19 +26,32 @@
 include "root.php";
 require_once "includes/config.php";
 require_once "includes/checkauth.php";
-if (ifgroup("admin") || ifgroup("superadmin")) {
-	//access granted
-}
-else {
-	echo "access denied";
-	exit;
-}
 
 
-$c = 0;
-$rowstyle["0"] = "rowstyle0";
-$rowstyle["1"] = "rowstyle1";
+//http get and set variables
+	if (strlen($_GET['url']) > 0) {
+		$url = $_GET['url'];
+	}
+	if (strlen($_GET['rows']) > 0) {
+		$rows = $_GET['rows'];
+	}
+	else {
+		$rows = 0;
+	}
 
+
+//set http compression
+	if (substr_count($_SERVER['HTTP_ACCEPT_ENCODING'], 'gzip')) {
+		ob_start("ob_gzhandler");
+	}
+	else{
+		ob_start();
+	}
+
+//define variables
+	$c = 0;
+	$rowstyle["0"] = "rowstyle0";
+	$rowstyle["1"] = "rowstyle1";
 
 //get the event socket information
 	if (strlen($_SESSION['event_socket_ip_address']) == 0) {
@@ -56,19 +69,20 @@ $rowstyle["1"] = "rowstyle1";
 		}
 	}
 
+//get information over event socket
+	$switch_cmd = 'show channels as xml';
+	$fp = event_socket_create($_SESSION['event_socket_ip_address'], $_SESSION['event_socket_port'], $_SESSION['event_socket_password']);
+	$xml_str = trim(event_socket_request($fp, 'api '.$switch_cmd));
+	//echo $xml_str;
 
-$switch_cmd = 'show channels as xml';
-$fp = event_socket_create($_SESSION['event_socket_ip_address'], $_SESSION['event_socket_port'], $_SESSION['event_socket_password']);
-$xml_str = trim(event_socket_request($fp, 'api '.$switch_cmd));
-//echo $xml_str;
-
-try {
-	$xml = new SimpleXMLElement($xml_str);
-}
-catch(Exception $e) {
-	//echo $e->getMessage();
-}
-//print_r($xml);
+//parse the xml
+	try {
+		$xml = new SimpleXMLElement($xml_str);
+	}
+	catch(Exception $e) {
+		//echo $e->getMessage();
+	}
+	//print_r($xml);
 
 //active channels array ----------------------
 	$channel_array = "";
@@ -92,7 +106,7 @@ catch(Exception $e) {
 	//echo "</pre>\n";
 
 
-// active extensions -----------------------------
+//active extensions -----------------------------
 	//get the extension information
 		if (count($_SESSION['extension_array']) == 0) {
 			$sql = "";
@@ -139,6 +153,15 @@ catch(Exception $e) {
 		//print_r($_SESSION['extension_array']);
 		//echo "</pre>\n";
 
+
+	//get a list of assigned extensions for this user
+		include "v_calls_active_assigned_extensions_inc.php";
+
+	//list all extensions
+		echo "<table width='100%' border='0' cellpadding='5' cellspacing='0'>\n";
+		echo "<tr>\n";
+		echo "<td valign='top'>\n";
+
 		echo "<table width='100%' border='0' cellpadding='0' cellspacing='0'>\n";
 		//echo "<tr>\n";
 		//echo "<td >\n";
@@ -148,21 +171,30 @@ catch(Exception $e) {
 		//echo "	&nbsp;\n";
 		//echo "</td>\n";
 		//echo "<td colspan='1' align='right'>\n";
-		echo "</tr>\n";
+		//echo "</tr>\n";
 
 		echo "<tr>\n";
-		echo "<th>Extension</th>\n";
+		echo "<th>Ext</th>\n";
 		echo "<th>Time</th>\n";
-		echo "<th>Direction</th>\n";
-		echo "<th>Profile</th>\n";
-		echo "<th>CID Name</th>\n";
-		echo "<th>CID Number</th>\n";
-		echo "<th>Dest</th>\n";
-		echo "<th>Application</th>\n";
-		echo "<th>Secure</th>\n";
+		if (ifgroup("admin") || ifgroup("superadmin")) {
+			if (strlen(($_GET['rows'])) == 0) {
+				echo "<th>Direction</th>\n";
+				echo "<th>Profile</th>\n";
+				echo "<th>CID Name</th>\n";
+				echo "<th>CID Number</th>\n";
+				echo "<th>Dest</th>\n";
+				echo "<th>App</th>\n";
+				echo "<th>Secure</th>\n";
+			}
+		}
 		echo "<th>Description</th>\n";
-		echo "<th>Options</th>\n";
+		if (ifgroup("admin") || ifgroup("superadmin")) {
+			if (strlen(($_GET['rows'])) == 0) {
+				echo "<th>Options</th>\n";
+			}
+		}
 		echo "</tr>\n";
+		$x = 1;
 		foreach ($_SESSION['extension_array'] as $row) {
 			//print_r($row);
 			$v_id = $row['v_id'];
@@ -225,54 +257,70 @@ catch(Exception $e) {
 				echo "<tr>\n";
 				echo "<td class='".$rowstyle[$c]."' $style_alternate>$extension</td>\n";
 				echo "<td class='".$rowstyle[$c]."' $style_alternate width='20px;'>".$call_length."</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>$direction</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>$sip_profile</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>$cid_name</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>$cid_num</td>\n";
+				if (ifgroup("admin") || ifgroup("superadmin")) {
+					if (strlen(($_GET['rows'])) == 0) {
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>$direction</td>\n";
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>$sip_profile</td>\n";
+						if (strlen($url) == 0) {
+							echo "<td class='".$rowstyle[$c]."' $style_alternate>".$cid_name."</td>\n";
+							echo "<td class='".$rowstyle[$c]."' $style_alternate>".$cid_num."</td>\n";
+						}
+						else {
+							echo "<td class='".$rowstyle[$c]."' $style_alternate><a href='".$url."cid_name=".$cid_name."&cid_num=".$cid_num."' style='color: #444444;' target='_blank'>".$cid_name."</a></td>\n";
+							echo "<td class='".$rowstyle[$c]."' $style_alternate><a href='".$url."cid_name=".$cid_name."&cid_num=".$cid_num."' style='color: #444444;' target='_blank'>".$cid_num."</a></td>\n";
+						}
+					}
+				}
 			}
 			else {
 				$style_alternate = "style=\"color: #444444; background-image: url('/images/background_cell_light.gif');\"";
 				echo "<tr>\n";
 				echo "<td class='".$rowstyle[$c]."' $style_alternate>$extension</td>\n";
 				echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
+				if (ifgroup("admin") || ifgroup("superadmin")) {
+					if (strlen(($_GET['rows'])) == 0) {
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>&nbsp;</td>\n";
+					}
+				}
 			}
+			if (ifgroup("admin") || ifgroup("superadmin")) {
+				if (strlen(($_GET['rows'])) == 0) {
+					if ($found_extension) {
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>\n";
+					}
+					else {
+						echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
+					}
+					echo "".$dest."<br />\n";
+					echo "</td>\n";
 
-			if ($found_extension) {
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>\n";
-			}
-			else {
-				echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
-			}
-			echo "".$dest."<br />\n";
-			echo "</td>\n";
+					if ($found_extension) {
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>\n";
+					}
+					else {
+						echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
+					}
+					if ($application == "fifo") {
+						echo "queue &nbsp;\n";
+					}
+					else {
+						echo $application." &nbsp;\n";
+					}
+					echo "</td>\n";
 
-			if ($found_extension) {
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>\n";
+					if ($found_extension) {
+						echo "<td class='".$rowstyle[$c]."' $style_alternate>\n";
+					}
+					else {
+						echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
+					}
+					echo "".$secure."<br />\n";
+					echo "</td>\n";
+				}
 			}
-			else {
-				echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
-			}
-			if ($application == "fifo") {
-				echo "queue &nbsp;\n";
-			}
-			else {
-				echo $application." &nbsp;\n";
-			}
-			echo "</td>\n";
-
-			if ($found_extension) {
-				echo "<td class='".$rowstyle[$c]."' $style_alternate>\n";
-			}
-			else {
-				echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
-			}
-			echo "".$secure."<br />\n";
-			echo "</td>\n";
-
 			if ($found_extension) {
 				echo "<td class='".$rowstyle[$c]."' $style_alternate>\n";
 			}
@@ -281,39 +329,41 @@ catch(Exception $e) {
 			}
 			echo "".$description."<br />\n";
 			echo "</td>\n";
+			if (ifgroup("admin") || ifgroup("superadmin")) {
+				if (strlen(($_GET['rows'])) == 0) {
+					if ($found_extension) {
+						echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
+							//transfer
+								//uuid_transfer c985c31b-7e5d-3844-8b3b-aa0835ff6db9 -bleg *9999 xml default
+								//document.getElementById('url').innerHTML='v_calls_exec.php?action=energy&direction=down&cmd='+prepare_cmd(escape('$uuid'));
+								echo "	<a href='javascript:void(0);' style='color: #444444;' onMouseover=\"document.getElementById('form_label').innerHTML='<strong>Transfer To</strong>';\" onclick=\"send_cmd('v_calls_exec.php?cmd='+get_transfer_cmd(escape('$uuid')));\">transfer</a>&nbsp;\n";
 
-			if ($found_extension) {
-				echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
-					//transfer
-						//uuid_transfer c985c31b-7e5d-3844-8b3b-aa0835ff6db9 -bleg *9999 xml default
-						//document.getElementById('url').innerHTML='v_calls_exec.php?action=energy&direction=down&cmd='+prepare_cmd(escape('$uuid'));
-						echo "	<a href='javascript:void(0);' style='color: #444444;' onMouseover=\"document.getElementById('form_label').innerHTML='<strong>Transfer To</strong>';\" onclick=\"send_cmd('v_calls_exec.php?cmd='+get_transfer_cmd(escape('$uuid')));\">transfer</a>&nbsp;\n";
+							//park
+								echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"send_cmd('v_calls_exec.php?cmd='+get_park_cmd(escape('$uuid')));\">park</a>&nbsp;\n";
 
-					//park
-						echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"send_cmd('v_calls_exec.php?cmd='+get_park_cmd(escape('$uuid')));\">park</a>&nbsp;\n";
+							//hangup
+								echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"confirm_response = confirm('Do you really want to hangup this call?');if (confirm_response){send_cmd('v_calls_exec.php?cmd=uuid_kill%20'+(escape('$uuid')));}\">hangup</a>&nbsp;\n";
 
-					//hangup
-						echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"confirm_response = confirm('Do you really want to hangup this call?');if (confirm_response){send_cmd('v_calls_exec.php?cmd=uuid_kill%20'+(escape('$uuid')));}\">hangup</a>&nbsp;\n";
+							//record start/stop
+								$tmp_file = $v_recordings_dir."/active_extensions_".$cid_num."_recording.wav";
+								if (file_exists($tmp_file)) {
+									//stop
+									echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"send_cmd('v_calls_exec.php?cmd='+get_record_cmd(escape('$uuid'), 'active_extensions_', escape('$cid_num'))+'&uuid='+escape('$uuid')+'&action=record&action2=stop&prefix=active_extensions_&name='+escape('$cid_num'));\">stop record</a>&nbsp;\n";
+								}
+								else {
+									//start
+									echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"send_cmd('v_calls_exec.php?cmd='+get_record_cmd(escape('$uuid'), 'active_extensions_', escape('$cid_num'))+'&uuid='+escape('$uuid')+'&action=record&action2=start&prefix=active_extensions_');\">start record</a>&nbsp;\n";
+								}
 
-					//record start/stop
-						$tmp_file = $v_recordings_dir."/active_extensions_".$cid_num."_recording.wav";
-						if (file_exists($tmp_file)) {
-							//stop
-							echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"send_cmd('v_calls_exec.php?cmd='+get_record_cmd(escape('$uuid'), 'active_extensions_', escape('$cid_num'))+'&uuid='+escape('$uuid')+'&action=record&action2=stop&prefix=active_extensions_&name='+escape('$cid_num'));\">stop record</a>&nbsp;\n";
-						}
-						else {
-							//start
-							echo "	<a href='javascript:void(0);' style='color: #444444;' onclick=\"send_cmd('v_calls_exec.php?cmd='+get_record_cmd(escape('$uuid'), 'active_extensions_', escape('$cid_num'))+'&uuid='+escape('$uuid')+'&action=record&action2=start&prefix=active_extensions_');\">start record</a>&nbsp;\n";
-						}
-
-
-					echo "	&nbsp;";
-				echo "</td>\n";
-			}
-			else {
-				echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
-				echo "	&nbsp;";
-				echo "</td>\n";
+							echo "	&nbsp;";
+						echo "</td>\n";
+					}
+					else {
+						echo "<td valign='top' class='".$rowstyle[$c]."' $style_alternate>\n";
+						echo "	2&nbsp;";
+						echo "</td>\n";
+					}
+				}
 			}
 
 			echo "</tr>\n";
@@ -340,6 +390,37 @@ catch(Exception $e) {
 			unset($write_rate);
 			unset($secure);
 
+			if ($x == $rows) {
+				$x = 0;
+				echo "</table>\n";
+
+				echo "</td>\n";
+				echo "<td valign='top'>\n";
+
+				echo "<table width='100%' border='0' cellpadding='0' cellspacing='0'>\n";
+				echo "<tr>\n";
+				echo "<th>Ext</th>\n";
+				echo "<th>Time</th>\n";
+				if (ifgroup("admin") || ifgroup("superadmin")) {
+					if (strlen(($_GET['rows'])) == 0) {
+						echo "<th>Direction</th>\n";
+						echo "<th>Profile</th>\n";
+						echo "<th>CID Name</th>\n";
+						echo "<th>CID Number</th>\n";
+						echo "<th>Dest</th>\n";
+						echo "<th>App</th>\n";
+						echo "<th>Secure</th>\n";
+					}
+				}
+				echo "<th>Description</th>\n";
+				if (ifgroup("admin") || ifgroup("superadmin")) {
+					if (strlen(($_GET['rows'])) == 0) {
+						echo "<th>Options</th>\n";
+					}
+				}
+				echo "</tr>\n";
+			}
+			$x++;
 			if ($c==0) { $c=1; } else { $c=0; }
 		}
